handle compilation package that do not have a folder
some packages (e.g. libstdc++) do not have their own package
(libstdc++ belongs to gcc), so the check on the existence of
the package name doesn't work, and was replaced with a mapping
between name and related package plus a file created after the
build succeed to avoid building it again.